rename argparsing to make resolving merge conflict easier
